pyproject.toml
[project]
name = "comfyui-automation"
version = "0.1.0"
requires-python = ">=3.10"
dependencies = [
    "comfyui-frontend-package",
    "comfyui-workflow-templates",
    "comfyui-embedded-docs",
    "torch",
    "torchsde",
    "torchvision",
    "torchaudio",
    "numpy",
    "einops",
    "transformers[timm]",
    "tokenizers",
    "sentencepiece",
    "safetensors",
    "aiohttp",
    "yarl",
    "pyyaml",
    "Pillow",
    "scipy",
    "tqdm",
    "psutil",
    "alembic",
    "SQLAlchemy",
    "av",
    "kornia",
    "mlflow",
    "spandrel",
    "pydantic",
    "pydantic-settings",
    "ftfy",
    "accelerate",
    "diffusers",
    "peft",
    "protobuf",
    "pyloudnorm",
    "gguf",
    "opencv-python",
    "pyjwt",
    "httpx",
    "websockets",
    "imageio-ffmpeg",
    "huggingface_hub",
    "ninja",
    "comfy-script[default]",
]
